Moved common code to MiscUtils
