+++ bottom-bar/src/main/java/com/roughike/bottombar/MiscUtils.java
@@ -20,7 +20,6 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewAnimationUtils;
 import android.view.ViewGroup;
-import android.widget.LinearLayout;
 import android.widget.PopupMenu;
 import android.widget.TextView;
 
@@ -119,6 +118,19 @@ class MiscUtils {
             }
         });
         animator.start();
+    }
+
+    protected static void resizePaddingTop(final View icon, int start, int end, long duration) {
+        ValueAnimator paddingAnimator = ValueAnimator.ofInt(start, end);
+        paddingAnimator.addUpdateListener(new ValueAnimator.AnimatorUpdateListener() {
+            @Override
+            public void onAnimationUpdate(ValueAnimator animation) {
+                icon.setPadding(icon.getPaddingLeft(), (Integer) animation.getAnimatedValue(),
+                    icon.getPaddingRight(), icon.getPaddingBottom());
+            }
+        });
+        paddingAnimator.setDuration(duration);
+        paddingAnimator.start();
     }
 
     /**
